requirements.txt
tk
Pillow
tensorflow
numpy
pandas
matplotlib
scikit-learn
opencv-python